@@ -1,11 +1,11 @@
 package rvega.bingo.presentacion;
 
 import rvega.bingo.negocio.MensajeApplication;
-import javax.enterprise.context.ApplicationScoped;
-import javax.faces.application.FacesMessage;
-import javax.faces.context.FacesContext;
-import javax.inject.Inject;
-import javax.inject.Named;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.faces.application.FacesMessage;
+import jakarta.faces.context.FacesContext;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
 import rvega.bingo.dominio.TipoModo;
 import rvega.bingo.socket.PushBean;
 
