@@ -1,9 +1,9 @@
 package rvega.bingo.presentacion;
 
 import java.util.logging.Logger;
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Inject;
-import javax.inject.Named;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
 import lombok.Data;
 import rvega.bingo.dominio.Usuario;
 import rvega.bingo.negocio.MensajeApplication;
