@@ -2,11 +2,11 @@ package rvega.bingo.presentacion;
 
 import rvega.bingo.negocio.MensajeApplication;
 import java.io.Serializable;
-import javax.enterprise.context.RequestScoped;
-import javax.faces.application.FacesMessage;
-import javax.faces.context.FacesContext;
-import javax.inject.Inject;
-import javax.inject.Named;
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.faces.application.FacesMessage;
+import jakarta.faces.context.FacesContext;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
 import lombok.Data;
 import rvega.bingo.dominio.Mensaje;
 import rvega.bingo.util.UsuarioSession;
