@@ -4,10 +4,10 @@ import rvega.bingo.util.CartonFactory;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
-import javax.annotation.PostConstruct;
-import javax.faces.view.ViewScoped;
-import javax.inject.Inject;
-import javax.inject.Named;
+import jakarta.annotation.PostConstruct;
+import jakarta.faces.view.ViewScoped;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
 import rvega.bingo.negocio.Bingo;
 import rvega.bingo.dominio.Usuario;
 import rvega.bingo.negocio.Rifa;
