@@ -1,8 +1,8 @@
 package rvega.bingo.socket;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.inject.Named;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.inject.Named;
 import org.omnifaces.cdi.Push;
 import org.omnifaces.cdi.PushContext;
 
