@@ -1,8 +1,8 @@
 package rvega.bingo.util;
 
-import javax.annotation.PostConstruct;
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Named;
+import jakarta.annotation.PostConstruct;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Named;
 import lombok.Data;
 import rvega.bingo.dominio.TipoModo;
 
