@@ -1,8 +1,8 @@
 package rvega.bingo.util;
 
 import java.io.Serializable;
-import javax.enterprise.context.SessionScoped;
-import javax.inject.Named;
+import jakarta.enterprise.context.SessionScoped;
+import jakarta.inject.Named;
 import lombok.Data;
 import rvega.bingo.dominio.Usuario;
 
